refactor: split views.py into focused modules with shared base class
Extract 18-view monolith (1556 lines) into four modules:
- base.py: BeatifyAdminView base class, RateLimitMixin, shared helpers
- game_views.py: game lifecycle views (start, end, rematch, gameplay, status)
- playlist_views.py: playlist CRUD and Spotify integration views
- stats_views.py: analytics, stats, and dashboard views
- views.py: HTML-serving views, core API views, and re-exports for backward compat

Closes #607
